+++ src/playlist-selectors.js
@@ -1,11 +1,11 @@
 import Config from './config';
 import Playlist from './playlist';
+import { parseCodecs } from './util/codecs.js';
 
 // takes seconds as input, outputs decay rate (unitless)
 const DEFAULT_EWMA_FAST_DECAY = decayRate(2);
 const DEFAULT_EWMA_SLOW_DECAY = decayRate(5);
 const DEFAULT_MOVING_AVERAGE_DECAY = decayRate(3);
-
 // Utilities
 
 /**
@@ -435,4 +435,36 @@ export const minRebufferMaxBandwidthPlaylistFilter = function(settings) {
   stableSort(rebufferingEstimates, (a, b) => a.rebufferingImpact - b.rebufferingImpact);
 
   return rebufferingEstimates[0] || null;
+};
+
+/**
+ * Chooses the appropriate media playlist, which in this case is the lowest bitrate
+ * one with video.  If no renditions with video exist, return the lowest audio rendition.
+ *
+ * Expects to be called within the context of an instance of HlsHandler
+ *
+ * @return {Object|null}
+ *         {Object} return.playlist
+ *         The lowest bitrate playlist that contains a video codec.  If no such rendition
+ *         exists pick the lowest audio rendition.
+ */
+export const lowestBitrateCompatibleVariantSelector = function() {
+  // filter out any playlists that have been excluded due to
+  // incompatible configurations or playback errors
+  const playlists = this.playlists.master.playlists.filter(Playlist.isEnabled);
+
+  // Sort ascending by bitrate
+  stableSort(playlists,
+    (a, b) => comparePlaylistBandwidth(a, b));
+
+  // Parse and assume that playlists with no video codec have no video
+  // (this is not necessarily true, although it is generally true).
+  //
+  // If an entire manifest has no valid videos everything will get filtered
+  // out.
+  const playlistsWithVideo = playlists.filter(
+    playlist => parseCodecs(playlist.attributes.CODECS).videoCodec
+  );
+
+  return playlistsWithVideo[0] || null;
 };