+++ src/vtt-segment-loader.js
@@ -279,6 +279,12 @@ export default class VTTSegmentLoader extends SegmentLoader {
 
     this.mediaSecondsLoaded += segment.duration;
 
+    if (segmentInfo.cues.length) {
+     // remove any overlapping cues to prevent doubling
+      this.remove(segmentInfo.cues[0].endTime,
+                  segmentInfo.cues[segmentInfo.cues.length - 1].endTime);
+    }
+
     segmentInfo.cues.forEach((cue) => {
       this.subtitlesTrack_.addCue(cue);
     });